Fix disconnect for latest `websocket_client`
In `websocket_client` versions 1.4.0 and later, the `ondisconnect/2`
callback is only triggered in the `connected` and `handshaking` states.
This causes the `tcp_closed` message to go to the `websocket_info/3`
callback instead of invoking the `ondisconnect/2` callback.

This commit updates the built-in transport to maintain consistent
behavior across the latest versions of `websocket_client`.